python/python-telegram-bot/requirements.txt
python-telegram-bot==20.7
feedparser==6.0.11
python-dotenv==1.0.1
